First commit of libdtrace-ctf.
This is the Solaris libctf, modified and extended for DTrace for Linux.
It is GPLed and publically available because it is used in the build
process of the Oracle Enterprise Linux kernel.  It's named as it is
because it is not expected to be useful outside of DTrace, and will
break its API and ABI as necessary to work with it.