Synch build system from DTrace.
This pulls in the symbol checking infrastructure (which libdtrace-ctf does not
yet use), but more importantly pulls in the -D_FILE_OFFSET_BITS=64 removal,
without which libdtrace-ctf would technically be binary-incompatible with
dtrace.

The incompatibility is not that important, since -D_FILE_OFFSET_BITS=64 happens
to have no effect on x86-64, but without this change libdtrace-ctf would be
being built with different CFLAGS when it was built independently of dtrace
than it was when it was built as a dtrace submodule, which seems to be a recipe
for disaster somewhere down the line.  So we synch the build system to
pre-emptively avoid any such disaster.