Stop parent types from overflowing into the child type ID space.
The libctf type space is 0xffff types long: its lower half is for parent types,
its upper half for child types, but this is not validated.  If you add more than
0x7fff types to a single CTF file, they silently overflow into the child space,
causing (at best) an assertion failure on the next ctf_update().  (If this
actually is a child CTF file, this will silently overwrite existing types
without warning.)

So add some overflow checking, similar to the checking that's already there to
prevent overflow of the ID space as a whole.