Prepare for varname->type-id tracking.
We need to track the mapping between variable names and type IDs in order for
DTrace to determine the types (as opposed to jut the addresses) of non-static
global variables in the kernel.  CTF already contains machinery to do just this,
but it operates in concert with an ELF string and symbol table, and we have
neither of these for the kernel.

So we need a new 'variable section' in libctf to track this.  Adding this will
break compatibility with Solaris libctf, since there is no way they will ever
add support for something like this in Solaris.  So since we're breaking
compatibility anyway, we change the CTF magic number, change the default section
name to '.dtrace_ctf', drop all support for older versions of CTF (version 2
plus variable offsets), reset the CTF version number to 1, and add support
for reading in and writing files with a header containing a variable-section-
offset field, though not for doing anything with the section this points to yet.