Allow changing the parent name of child CTF containers.
This requires ripping out the old code which hardwired a name of "PARENT" and
introducing a new dynamically-allocated field in the ctf_file_t to store a
parent when it is set dynamically.  It is injected into the first element of the
string table at ctf_update() time (just like "PARENT" used to be) and the string
table offsets adjusted accordingly.  If no parent name is set when ctf_import()
is called, it is set to "PARENT" just as before, but the caller can reset it
freely.  (All this applies only to read-write containers. You can reset the
parent name of read-only containers, too, but since you cannot call ctf_update()
on such containers it is not much use.)

This is a precondition for any future parallelization of dwarf2ctf, and allows
some ugly code hardwiring the name of dtrace_ctf as the parent of all kernel
modules' CTF sections to be removed from DTrace userspace.