Install libraries to lib64, not lib.
(This actually varies per distro, but most distros use lib64 -- and most that
used lib32 now use multiarch, which we make no attempt to support here.
It also varies per bitness, but we make no claims that this works, or is
useful, on non-64-bit platforms.  It certainly is not tested there, hence
the trap at the top of the makefile.)