Use a better hash function for the dynamic type definition hash table.
The hash function this used to use was type & (buckets - 1), which is not
particularly good if for whatever reason your bucket count is not a power of two
-- and if it is a power of two it's just the same as %. Sure, & is faster, but
this is premature optimization: there are many larger inefficiencies in libctf
than this, and if it's slower it's *vastly* slower, often piling absolutely
everything into one bucket.

Meanwhile, because ID numbers are consecutively assigned, % gives a perfect
spread, just as & was intended to.