Fix CTF type comparison.
libctf's ctf_type_compat() function claimed to determine if two types were
assignment-compatible, and our DTrace relies on this, since users will tend to
type in things like 'unsigned long', but our DWARF output names that type
'unsigned long int', and we haven't injected typedefs for this sort of thing
unless the DWARF already contained them.

Unfortunately, its definition of assignment-compatibility was very different
from C's, defining two types as assignment-compatible if they had identical
names *and* structure / encoding / whatever.  This is very far from the C
definition.  While DTrace actually relies on the 'same name' property
for structures, it does not rely on it for integers, so we change the
definition to

 - same type, or
 - same encoding if base type (int, float et al), or
 - compatible base type if pointer, or
 - compatible base type, element count and index type if array, or
 - same names if enum or forward, or
 - same names and member count if structure

This should suffice.