Introduce new ctf_variable_iter() iteration function.
This iterates across all variables in a given file, passing the variable name
and associated type ID of each to a callback.  Iteration fails immediately if
this is a child CTF file with no associated parent.

Just as for the other iteration functions, the CTF file ID is not passed, so if
you want to use the type ID in the callback, you have to either pass it in a
global or in the state argument.

Add a new symbol version containing ctf_variable_iter(), and bump the shlib to
libdtrace-ctf.so.1.1.0.