Do not look off the end of allocated strings when looking up types by name.
ctf_lookup_by_name() endeavours to eliminate type prefixes (such as 'struct') by
matching each such prefix and walking past it.  Unfortunately the code that does
so determines compares the two strings to see if it can skip over the prefix
using strncmp().  This is not safe nor correct: a strncmp() of two bytes against
the prefix "struct" when looking at a variable named "st", or of zero bytes when
the string is empty, will return zero, but that doesn't mean it is safe to skip
six bytes to jump past the word "struct", and it certainly doesn't mean that the
prefix *is* 'struct'.

So only consider the comparison successful if strncmp() returned 0 *and* the
string being compared is at least as long as the prefix.  This preserves the
intended behaviour of comparing only the prefix without assuming that the prefix
is there when it is not.