Add ctf_snapshot() and ctf_rollback().
ctf_snapshot() returns an ID (actually a structure) which can be
passed to ctf_rollback() to throw away all types and variables added
since the snapshot.  This is just like ctf_discard() was with respect
to ctf_update(), except that ctf_snapshot() is many orders of
magnitude cheaper, and you can call ctf_snapshot() many times and
roll back to any of them.

You cannot roll back to a snapshot taken before the last ctf_update()
call: the new error ECTF_OVERROLLBACK is returned in this case.  (A
ctf_discard() might suffice to discard some of the added types.)

The old ctf_discard() is reimplemented in terms of ctf_snapshot(),
and the old machinery which served to track variable creation times
in terms of the number of ctf_update() calls invoked since ctf file
creation is now re-expressed in terms of the number of ctf_update()
plus ctf_snapshot() calls.