Fix bitfields.
Bitfields are somewhat unusual monsters in CTF: they are described
not purely a type but by a union of (type, encoding), and are usually
non-root types.  You may perfectly well find a non-root type representing
a bitfield followed by an identical root type representing a
non-bitfield of the same width.  libdtrace-ctf wasn't handling this at
all.  Nor could it handle the case where one of the types exists in the
committed structures and one in the pending list used before
ctf_update() is called.

Add some extra checking for multiple root-visible types with the same
name: previously, this just led to a random one being returned, but now
it can actually lead to a valid type being replaced with an invalid one
with a bitness too small to represent its contents.

(Lightly modified version of pre-existing Solaris patch from Robert
M. Harris.)