Avoid erroring on root type conflicts.
We have root type conflicts in the core kernel: the previous behaviour
always allows this type to be overridden and always picks the last type
on use, which leads to the first type (which is hilariously corrupted)
causing no harm at all.  Revert to this behaviour, avoiding the conflict
return, so that dtrace running with a new libdtrace-ctf and a pre-UEK5
kernel does not cause spurious conflicts.