create: add ctf_compress_write()
libdtrace-ctf has long contained machinery for reading in compressed
archives, which are structured as a CTF header with the CTF_F_COMPRESS
flag set, followed by a zlib-compressed data stream.  However, there is
no machinery to write such a stream out, which is unfortunate because
it is better in many ways than the 'just compress the whole thing
with gzwrite()' approach used by dwarf2ctf at present: principally, the
header is uncompressed, so parent labels and in particular the file
length can be determined without uncompressing anything.  This means
that CTF_F_COMPRESSed files are usefully concatenatable: the individual
streams can be identified for reading without uncompressing the entire
stream first.

We'll need this property for CTF archives, so add a new API,
ctf_compress_write() which writes out the passed-in CTF file to the
given file descriptor.

Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>