Introduce CTF_VERSION_2, with more types, structure members, and kinds
CTF format 1 is showing its age.  It supports only 2^15-1 types in any
one container, which is little enough that a make allyesconfig or
allmodconfig kernel will fail to build because it runs out of types
(each enumeration value counts as a type!).  It supports only 1024
structure members or enumeration values in any one struct/union/enum,
which is little enough that some machine-generated code in the kernel is
already hitting it; and it supports only 32 type kinds, which we are
halfway to using up already.  (Yes, the minimum C standard limit is
similar, but almost all compilers support many more, in particular
including GCC and all compilers that have a hope of compiling the Linux
kernel -- or the Solaris kernel, come to that.)

All of these problems arise because the structures used to encode types
internally are too small.  We can boost all of these limits by making
three changes:

 - make types a uint32_t, allowing 31 bits for types in each of the
   parent and child containers (and incidentally allowing much larger
   sizes to be encoded in the "small" ctf_stype and ctf_member
   structures).  This causes the ctf_stype, ctf_type, ctf_array,
   ctf_member and ctf_lmember structures to grow, but we have been
   able to drop padding so ctf_lmember has grown less than might be
   expected.

 - make the info word a uint32_t too, allowing 32 bits for the info
   string, permitting 16 million structure/union/enum members rather
   than 1024, and while we're at it adding a bit to the type kind so
   we now support 64 of those.  (This uses no extra space because the
   space it consumes was occupied by alignment padding in v1.)

Backward-compatibility here is a bit tricky.  The ABI and soname are
unchanged, but there are a lot of macros in the installed <sys/ctf.h>
that provide direct access to the guts of CTF files.  None of these are
used by DTrace or dwarf2ctf, so I have freely changed most of their
names from e.g. CTF_TYPE_* to CTF_V1/V2_TYPE_* to force affected users
to recompile and rethink what they're doing: everything whose name is
unchanged has unchanged semantics, but may now refer to V2 rather than
V1 (particularly true of limit constants and structure names, where the
unadorned name is v2 and the name suffixed _V1 or _v1 is for the older
format).  Some things that were macros have had to become functions:
CTF_TYPE_ISPARENT and CTF_TYPE_ISCHILD are now API functions, named
ctf_type_isparent() and ctf_type_ischild().  The _V1-named macros are
explicitly to access CTF files by hand: types returned from the API will
always appear to be v2, even if the underlying CTF file was v1.

We handle backward-compatibility concerns by transparently upgrading v1
CTF files to v2 at ctf_bufopen() time, via a new upgrade_types().  This
upgrades the entire type section, including varlen regions, which is
quite expensive but is easy to validate and only applies to old CTF
files in any case.  symsects are not updated and cause an open failure,
because dwarf2ctf doesn't create them so there should be zero instances
of this in the wild.  There are a number of assertions in the
transparent upgrading process to ensure that the entire region is
converted and that various invariants that might cause buffer overruns
if they go wrong are not violated.  (These invariants should be violated
only if the code is buggy: corrupted CTF files should not be able to
cause them, though the existing code is not at all armoured against
malicious CTF files in any case.)

(The value of the CTF_VERSION_2 constant is 3 rather than 2 to allow
room for a use of 2 in a later patch in this series.)

Other internal-only changes necessary for this:
 - The ctf_fileops vector has grown new members allowing the
   ctf_get_ctt_size() to vary depending on CTF file version
   (transparently upgraded to v2 after open is complete), and
   allowing the manual computation of vbytes repeated over and over
   again in ctf_open to be moved into a single LCTF_VBYTES() macro.
 - Fairly large chunks of ctf_bufopen() to do with updating the ctf base
   and version number have moved into functions so that upgrade_types()
   can call them as well: a bunch of code that depends on regions that
   transparent upgrade may reallocate has been moved below init_types()
   as well.
 - CTF_INFO_KIND(), CTF_INFO_ISROOT() and CTF_INFO_VLEN() have become
   LCTF operations that take a CTF file as a parameter, to allow them to
   dynamically adjust to v1 versus v2 (though in practice most calls
   will always be v2 because of transparent upgrade, if we should be
   wrong about this the code will now adapt).
 - CTF_INDEX_TO_TYPE and CTF_TYPE_TO_INDEX have become LCTF_ operations
   that take a CTF file as a parameter.  This is because while we
   upgrade the file format, we do not change the type IDs, so the old
   parent/child boundary of 0x7fff persists in a v1 upgraded to v2.
   We record this in a new member fp->ctf_parmax, and compare against
   this member when determining whether a type is a parent or a child.
 - Many casts of type IDs to ushort_t have become casts to uint32_t
   instead.
 - A few out-of-memory errors have been flipped to ENOMEM from EAGAIN
   for consistency.  It is unlikely anyone will be affected.

The library is (and was, and remains) rather confused about whether it
wants to use a uint_t or a uint32_t to encode things, and continues to
rely on no information loss happening when converting in either
direction.  We enforce that their sizes are identical with a C11 static
assertion (if supported by the compiler, which presently requires overriding
CFLAGS at build time).  A later commit in this series converts the lot
to uint32_t.

All of this has performance impact well below the noise floor, at least
on a machine big enough to keep the CTF file in L3 cache across ctf_bufopen().
If one times dtrace make check-stress runs using the old library, the
new library using CTF generated by the old library, and the new library,
one sees something like this (these figures generated in a QEMU VM atop
a mostly-idle Xeon E5-2640, i.e. it had 24MiB of L3 cache to play with:
machines with much less cache might be more impacted):

old:
1125.29user 677.44system 51:46.49elapsed 58%CPU (0avgtext+0avgdata 91712maxresident)k
132900inputs+4509664outputs (43major+48812910minor)pagefaults 0swaps

new, old ctfa:
1128.35user 686.25system 51:38.97elapsed 58%CPU (0avgtext+0avgdata 91648maxresident)k
15484inputs+4541536outputs (0major+48395888minor)pagefaults 0swaps

new, new ctfa:
1126.85user 696.26system 51:55.69elapsed 58%CPU (0avgtext+0avgdata 91684maxresident)k
254756inputs+4570376outputs (136major+47324039minor)pagefaults 0swaps

The range expansion does lead to CTF files growing larger:
-rw-rw-r--.  1 oranix oranix    6311731 Jun 13 14:42 vmlinux.ctfa
-rw-rw-r--.  1 oranix oranix    5950771 Jun 13 12:16 vmlinux.old.ctfa

However, this is no more than the growth we often see between kernel releases
anyway.  The files which account for the majority of the types (vmlinux.ctf
and shared_ctf.ctf) have grown by less than 64KiB each.  Large amounts of
space could be saved in a later format revision by exploiting knowledge of
the textual structure of C identifiers (accumulating underscore-separated
pieces in a table and composing the identifiers out of the pieces).  But
that is for format v3. :)

Orabug: 28150489
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>