Boost the hash table sizes again
With the support for many more types, it is now common to have 20,000+
types in a single container.  Boost the hash bucket count to 8191 to
allow for this.  (The next upgrade should be to dynamic sizes, because
*no* static size can suffice for the sheer number of distinct types we
can now encode.)

A little trivial testing suggests that before this change, the average
length of the hash chains searched by a given successful lookup when
running a full DTrace testsuite run was 9: afterwards, it was 3 (so
the average chain length was probably double that). (This is only an
average: some are much worse, with averages of >30 visible on some
hashes before the change: still better than a linear search but
that's about all you can say for it.)

No effect on speed or memory usage above the noise is visible, despite
this: it probably takes contrived testcases to see one, or a machine
with less L3 cache than the one I'm testing on.

Orabug: 28150489
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>