Introduce ctf_getdatasect().
This returns the ctfsect passed to ctf_open(), so the caller doesn't
need to keep track of it.  We are careful never to free this, but have
before now required the caller to keep hold of it so it could free it
itself.  This is needlessly clumsy, when we can hand it back: the caller
can ask for it right before ctf_close(), then free the data section it
is passed.  (It is always safe to do so, since we never free it.)

Thanks to Tomas Jedlicka for the suggestion.

Orabug: 28150489
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>