uint32ify the API.
We already depend on uint_t and uint32_t being the same size, so just
migrate to using uint32_t everywhere.  The uint_t type is still
provided for any other users that may rely on it, but not used.

Because type sizes have not changed, this technical API and ABI
change is not an actual change on any platform the library is
likely to run on, so the soname is unchanged with this release.

Suggested by Tomas Jedlicka.

Orabug: 28150489
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>