ctf_ar: only upgrade and recompress on request
Currently, ctf_ar extracts archives by opening each CTF file in turn,
then writing out its compressed representation.  This is not very
desirable default behaviour for an archiving tool: by default, such
tools should write out precisely what it is in the archive, which for
ctfa archives means the data will sometimes be uncompressed (if it is
below the threshold specified to ctf_arc_write() at write time) and will
be v1 archives if the origin was v1, rather than being v1-upgraded-to-v2
hybrids.

We add a new ctf_archive_raw_iter() function which allows iteration over
the raw contents of an archive, and use it in ctf_ar by default, unless
the -u option (triggering dynamic upgrading) is provided.

(We retain the option to rewrite them because it is a very useful bulk
tester for the transparent upgrade code and for the upgrade-and-write-
out-again path.)

Orabug: 28150489
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>