ctf_add_type: Add back some conflict workarounds.
Commit c4588475f94bff856ca2edfb3be82c1b648c9ad4 started checking that
multiple root-visible types with the same name but different definitions
did not exist.  Unfortunately, dwarf2ctf had a bug which led to two
root-visible definitions of int existing, one only four bits wide
(mistakenly using a count of bytes, not bits, in one place): in even
earlier releases it was only one bit wide.  So we had to rip this
checking out in release branches almost at once because the
ctf_add_type() of the core D int type failed due to conflict with this
ridiculous erroneous int type.

Eventually we stopped doing so, and nothing went wrong -- but now people
are testing new libdtrace-ctf with older kernels built with the buggy
dwarf2ctf and things are failing again.

So introduce a narrower check which can be reasonably carried along
indefinitely: only and specifically allow root-visible conflicts with 1-
and 4-bit integer types named 'int'.  This means that the DTrace int
type will override the buggy definition for most uses and things will
keep working, while not losing the vast majority of the protection for
the rest of us.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Orabug: 28952429