ctf_bufopen: do not mprotect() the heap if reading from ctfsect
We try to mprotect() the CTF buffer data we allocated after we finished
footling around with it (opening it, upgrading it, whatever) to prevent
later overwrites from smashing it accidentally.  This is usually safe
because we are careful to always allocate CTF data using
ctf_data_alloc(), which translates to mmap().

Unfortunately, we failed to cater for the case where the CTF data has
been passed in from a CTF section allocated by someone else: in this
case, if it doesn't need to be reallocated by the upgrade code, we
end up mprotect()ing part of the heap, leading to intermittent, bizarre,
situation-dependent crashes depending on what else happens to have been
allocated in the same memory page.

Do not mprotect() in that case.

Reported-by: Simon Xu <simon.xu@oracle.com> # much thanks!
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Orabug: 28952429