ctf_create: add functions for explicitly-sized struct/union creation
This lets callers create structs and unions with sizes explicitly
specified by the DWARF, rather than always relying on the required
padding being the same as that implied by the alignment of the last
member.

If you add so many members that the size exceeds this figure, we revert
automatically to the previous behavour (so passing in a size of 0 to
ctf_add_struct_sized() or ctf_add_union_sized() is equivalent to calling
the pre-existing ctf_add_struct()/ctf_add_union() functions.)

This is an ABI addition: bump the soname to 1.5.0 and update the symbol
versions accordingly.

Needed by an upcoming dwarf2ctf fix.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Orabug: 29054972