GNUization: file renamings to GNU style.
The renaming here is a little strange, for compatibility reasons:
in the source tree (and in the Cygnus tree), -'s are used to separate
words in filenames, but in the installed tree include/sys/ctf-api.h is
renamed to sys/ctf_api.h, so out-of-package userspace need not change.
ctf_types.h, being a compatibility header unused by libctf and not
installed into the Cygnus tree, does not change its name at all.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>