Cross-endianness support.
This is implemented much like transparent type upgrades for V1->V2:
if the endianness is visibly wrong, we work across the header and all
sections that contain things that might need swapping, and swap them.

We use the <byteswap.h> routines in conjunction with a bit of sizeof()
to automatically swap each item correctly without needing to manually
get type sizes right all over the place.  We use a C11 static assertion
to ensure that we never try to swap something whose size we cannot
handle.  (We switch from gnu99 to gnu11 -std mode in consequence.)

Tested by generating CTF for a fairly-recent still-working Solaris
UEK kernel (UEK4-QU5), and comparing the results of a ctf_dump of
every CTF file in the resulting .ctfa with a dump of the same .ctfa
on x86-64.  (No differences.)

Sections not used by dwarf2ctf (labels, data objects, and function info)
are untested, but these are also the simplest sections and if they don't
work will be trivial to fix.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>