hash: rename to allow for later hashing things other than helems
Right now the ctf-hash.c implementation allows hashing only
ctf_helem_t's, so it is completely specialized for hashing in type IDs,
and indeed only type IDs in the non-dynamic table.

This is ridiculous: we have multiple hashtables in libctf alone which do
not conform to this constraint, so we end up exporting most of
ctf-hash's guts and using them to implement multiple even *worse*
hashtables by hand in ctf-create.c.  But we can't fix this until we
rename the hashing functions that take and return specifically helems
so that there is room in the namespace for more general functions.

Rename things accordingly.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>