hash: replace the dynamic hash implementation with GHashTable
libctf contains not one, not two, but three hashtable implementations:

 - one in ctf-hash.c, which is specialized for hashes from string (as a
   uint32_t strtab pointer) -> ctf_helem_t, with a fixed-at-creation-
   time count of elements (as many elements as there are types,
   roughly), used for the frozen type tables used post-ctf_update()
   alone: this is the only table used by read-only CTF files.  The only
   element of the ctf_helem_t that any of the callers actually use is
   the h_type element, so this could perfectly well be a hash from
   string -> type ID.  It has a fixed count of buckets, no matter the
   number of types.

 - two in ctf-create.c for hashing dtdefs and dvdefs during creation,
   fixed in size to 1024 elements.  These are both hashes from
   type ID to struct pointers.

They all use the same not-terribly-good hash function, but they have
completely divergent implementations otherwise.  The ctf-create pair, in
particular, really should be dynamically-sized, and should as much as
possible use the same interface as the ctf-hash one.  The ctf-hash one
should not have a fixed bucket count, but should (like libiberty) use
the lowest power-of-two prime higher than the count of elements.

We fix both of these by adding a new hash type which is implemented as a
GHashTable. From outside ctf-hash.c this appears to be a typedef to a
struct ctf_dynhash, but this structure does not exist inside ctf-hash.c:
instead, we cast to GHashTable wherever needed.  This makes it easier to
replace ctf_hash_t with a completely different implementation, without
needing any extra layers of indirection or memory allocation (and with
only a few extra casts needed).  We use a slightly different hash
function for integers than the default GHashTable one, because that
default is great for pointers and fairly dreadful for integers
distribution-wise.  Simply multiplying by a fixed small prime fixes that
completely (already done in upstream glib 2.60.0+, but we can't rely on
a glib that recent!)

ctf_hash remains unchanged, except that the hash function is the same as
used for ctf_dynhash (it's slightly better than the one we used before)
and all the types are fixed (no more unsigned long / unsigned short).

Changes visible outside ctf_hash.c:

New ctf_dynhash_*() functions, and customizable hashing/
equality-checking functions passed to ctf_dynhash_create().
(ctf_hash_create() accepts these too for ease of replacement of both
hashes with one underlying implementation, but is allowed to fail if
these are not ctf_hash_string()/ctf_hash_eq_string().)

ctf_hash_create() now returns a newly-allocated hash, or NULL.
ctf_hash_lookup_type() loses its length parameter and now returns a
ctf_id_t rather than a ctf_helem_t.  The hardwired hashlens in
ctf_file_t and ctf_hash_compute() are gone.  We now have a prime number
table (generated using the code from libiberty) which we search at
startup: doing this rather than our previous fixed-size buckets reduces
the average chain length in a test vmlinux.ctf from 2.6 to a little over
1.2.

Other than a little restructuring to eliminate the need for the helem_t
to be publically visible, this is all very easy to adapt to. dwarf2ctf
speeds up by 25s from this change alone :)

No change to read-time (e.g. core DTrace usage) is visible above the
noise floor, though it is possible to imagine improvements via
e.g. using a hashtable implementation that allows one-off space
reservation (as ctf_hash.c used to do), or adding such a feature to
GHashTable, or writing out a perfect hash or a description of a suitable
perfect hash function and using that without needing to do any
population at all... but this will do for now.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>