TODO: mark two as done, drop one as faulty
We've moved to a decent hash: we don't want to drop ctf_txlate because
despite appearances from a quick grep it is in fact used, and crucial,
being the primary mapping from type ID to type packed in the ctf_buf
(see uses of 'xp' in init_types()).

Note the endianness changes, too, for completeness.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>