create: reduce the need for ctf_update()
The rule when creating types using the ctf_add_*() API functions
(henceforth "adders") is that whenever you pass some pre-existing type
ID into an adder that takes one (say, ctf_add_member_offset(), where you
have to pass in a pre-existing type for the member), you must have
called ctf_update() between the addition of that pre-existing type and
the addition of this one.

This is done so that type accessors in the implementation of ctf_add_*()
do not need two codepaths for types in the static space and
recently-added dynamic type definitions.  This means that large-scale
type creation tools cannot just call ctf_add_*() repeatedly and then do
a single ctf_update() at the end, but need to call it over and over:
even with optimizations to call ctf_update() only when a type addition
actually fails with ECTF_BADID, dwarf2ctf calls ctf_update() about ten
thousand times.  This could be reduced by ranking the types so that all
types that other types depended on were added before any of the types
that depended on them, but wouldn't it be easier to just fix things so
this nonsense wasn't needed?

This would be moot if ctf_update() were cheap, but alas ctf_update() is
monstrously inefficient.  It reconstructs the entire type array and
builds a new ctf_file_t using ctf_bufopen(), then swaps the internals of
this with the internals of the old one and frees the "new" (== old) one,
keeping around only the dynamic state which is not carried in the static
CTF buffers: this involves multiple traversals of the DTDs to build the
static buffers and then multiple more traversals to parse it at open
time.  This means that for any reasonable number of types it blows at
least the L1 and L2 caches repeatedly every time you call it, and does
some massive iterations and a bunch of big memory allocations.  Doing
this more often than we must is bad.

Thankfully, now that we have a proper dynhash abstraction, fixing this
turns out to be quite easy.  Despite the warning above ctf_update(), the
type addition functions don't actually require many other bits of libctf
to become aware of the dynamic type definitions.  Only those places that
are called from ctf_add_*() functions and need to access the list of
members, array types or encoding need changing: these things normally go
into vlen data, which isn't constructed until ctf_update() is called, so
anything that accesses them at this stage needs to hunt them down in
ctf_dtdef_t.dtd_u instead, where the adders put them.

There are four such places:

 - ctf_lookup_by_id() needs to know if a type is dynamic to know if it
   should dig the ctf_type_t out of the DTD's dtd_data member
 - ctf_type_align() needs to figure out the alignment from a dynamic
   structure/union's dtu_members list
 - ctf_type_encoding() needs to fish out the encoding from dtu_enc
 - ctf_array_info() needs to fish out the array info from dtu_arr

There is less change here than you might expect because most of these
functions are actually API functions, so the fact that the dynamic info
is stored in the same data structures as the API uses to pass things
back means we don't need to massage the data once we fish it out: we can
just pass it back.  ctf_type_align() is more voluminous, but only
because we need to change a level of indentation.

We add one function which all the above call, ctf_dynamic_type(), to
determine whether a type is dynamic or not, and fish out the DTD if it
is.

The adders themselves also need small changes.  The struct/union/ enum
adders already identify CTF_K_FORWARDs in the static space and promote
them appropriately (so that "struct foo;" can be turned into "struct
foo" with members, but not into "union foo"): we need to do that if
these types are dynamic, without requiring us to make four times as many
hashes to distinguish the struct/union/enum namespaces the way the core
static hashes do.  (That code is repetitive enough as it is).  To do
this, we need to be able to look up these types by name, but ctf_dthash
is indexed by type ID, so we can't use that.

Similarly, the forward adder also needs to identify an existing forward
and just return it, since redeclaration of forwards via e.g.
"struct foo; struct foo;" is perfectly valid.

So introduce a new dynhash, ctf_file_t->ctf_dtbyname, which maps
identifiers to DTDs just as ctf_dthash maps ctf_ids to DTDs.  To
separate the namespaces, we prefix the identifiers with "struct ",
"union " or "enum ", producing identifiers that are invalid in C and
thus cannot collide with any user-introduced identifiers.  (Because
these are prefixed, they are newly-allocated: use the new dynhash member
freeing machinery to free them automatically).  This is not costly
because allocations are quite rare: only one per newly-added
structure/union/enum, not one per added member.

On the userspace API side, to make it possible for callers to tell when
they can avoid calling ctf_update(), we introduce a new compatibility
define, LIBDTRACE_CTF_OMISSIBLE_CTF_UPDATE, which if set indicates that
ctf_update() need not be called if all you're calling is adders.  (You
still need to call ctf_update() if you're going to call functions other
than ctf_add_*() / ctf_discard() / ctf_snapshot() yourself.)

No change to generated types that isn't accounted for by long-standing
bugs elsewhere (fixed in later commits).  dwarf2ctf runtime drops by a
wildly variable amount never seen to be less than 60s (30% of total
runtime) and sometimes as high as 50% (I suspect the variance is down to
cache colouring effects, since we are doing so many huge traversals that
cache timings will dominate its running time, which is much more
variable before this change than afterwards).

XXX fix those timngs given true dwarf2ctf runtime, not make ctf

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>