ctf: drop the horrible ctt_type / ctt_size macros when __GNUC__
I just wasted most of an hour hunting down a bizarre compilation error
that happened because I named a variable 'ctt_size' in code that
included <ctf.h>.

This is because ctf.h implements unnamed union fields via #defines
and named unions: but the #define's name is quite possible to use in
userspace code, and pollutes the namespace needlessly.

We can't avoid it with other compilers, but on GNU C we can rely on
unnamed union fields freely.

(There is still one instance of this in <ctf.h>, but we can't make the
cth_preamble an unnamed struct because it is directly referenced by
name.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>