types: introduce ctf_type_aname for dynamically-allocated type names
The type-name returning functions ctf_type_name() and ctf_type_lname()
both work by taking a buffer and a size and returning as much type name
as will fit into the buffer and perhaps an overflow error: the only
difference between ctf_type_name() and ctf_type_lname() is that the
latter returns an error on truncation.

These are both deeply clumsy to use if you want an actual untruncated
value: you need to call it repeatedly, possibly allocating a bigger
buffer in between.  Instead of all of this, implement the two in terms
of a new ctf_type_aname() that just returns a dynamically-allocated
string, and change the underlying ctf-decl.c routines to always use
dynamic allocation and dynamic string appending.

ctf_type_name() and ctf_type_lname() remain, but ctf_type_aname() is
recommended from now on.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>