lib: tweak to not require code changes when the CTF version changes
The ctf_version() function, meant to allow transparent changes of CTF
version on the fly, currently has a check to ensure that nobody tries to
use it for that (since this has not been implemented, at least not yet).

The check hardwires in the latest CTF version number, so it is easy to
overlook whenever this is bumped: this then fails at compile time in
a NO_COMPAT tree, because version constants before the latest do not
exist there, and checks for the wrong thing everywhere else.

Conditionalize on CTF_VERSION, which matches the latest version
(whatever it is), instead, to suppress this.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>