open: introduce ctf_simple_open
This is like ctf_bufopen but takes the component parts of the ctf_sect_t
that ctf_bufopen actually uses (not many of them) rather than a
ctf_sect_t.  (It calls ctf_bufopen internally.)

This means that its prototype is necessarily less stable than the
others: if ctf_bufopen() comes to need more pieces of the ctf_sect_t, it
will need to grow more parameters -- but until that time, it is much
simpler to call in the common case that you don't actually have a
ctf_sect_t yourself so are forced to mock up a fake one just to placate
ctf_bufopen (and then you are encoding the information about which
elements of the ctf_sect_t are used and which are not in a random
structure initialization, and the compiler will not notice if that
assumption becomes invalid the way it will if a function parameter is
missing).

Half of libdtrace-ctf's own uses of ctf_bufopen() are in this boat, as
are all the uses in binutils.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>